pyproject.toml
[tool.poetry]
name = "clockify_pipeline"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
requests = "^2.32.3"
minio = "^7.2.15"
boto3 = "^1.38.12"
pandas = "^2.2.3"
pyspark = "3.5.0"
jupyterlab = "^4.4.2"
apache-airflow-providers-apache-spark = "^5.2.1"
apache-airflow = "2.10.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
